Add files via upload
Library Management System is designed for managing and storing all the records of Books in the Library. Once the User opens their account then he is eligible to access many more features available in Library Management System, user can see all books available, user can issue a book, and submit the book, If the user does not submit the book within the given days he has to pay fine for no of days late he submits the book. Also, one more condition is that the user can issue three books at a time to issue more books he first 
needs to submit the book. If the user lost the book then he will pay the fine 
as the price of the book.